Compiles but has runtime errors due to Guava not being GWT Serializable

+++ webprotege-gwt-ui-client/src/main/java/edu/stanford/bmir/protege/web/client/pagination/PaginatorViewImpl.java
@@ -114,7 +114,7 @@ public class PaginatorViewImpl extends Composite implements PaginatorView {
 
     @Override
     public void setElementCount(long elementCount) {
-        String msg = messages.pagination_rows(elementCount);
+        String msg = messages.pagination_rows((int) elementCount);
         rowCountField.setText(msg);
     }
 }
